Added a new host plugin to update quota and avoid aggregates
Added a new host plugin which is very similar to the existing
plugin but has patches where nova-aggregates are being created.

Since there is no need to create and maintain aggregates, the
corresponding code has been commented. Also, added code to update
quota for the project on lease_start and lease_end events.

Change-Id: I7f1562d7b531e600bbd3896911d92fc1d3df2057